% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctocntry.r
\name{cctoname}
\alias{cctoname}
\title{UN cc to country name}
\usage{
cctoname(cc, wpp = "wpp2019")
}
\arguments{
\item{cc}{UN country code}

\item{wpp}{which WPP to use}
}
\value{
country name
}
\description{
This function allows you to extract country name for give UN country code
}
\examples{
cctoname(900, "wpp2019")
}
\keyword{read}
