% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nametocc.r
\name{cntocc}
\alias{cntocc}
\title{UN country name to country code}
\usage{
cntocc(cn, wpp = "wpp2019")
}
\arguments{
\item{cn}{UN country code}

\item{wpp}{which WPP to use}
}
\value{
country code
}
\description{
This function allows you to extract country code for give UN country name
}
\examples{
cntocc("World", "wpp2019")
}
\keyword{read}
