% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapr.r
\name{eapr}
\alias{eapr}
\title{Convert given edu distribution into eapr}
\usage{
eapr(prop, edutcol = F)
}
\arguments{
\item{prop}{distribution data 'edu' (sorted) 'value' as proportion for single age-group}
}
\value{
eapr datatable
}
\description{
This function returns EAPR for given education distribution
}
\keyword{read}
