% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapr1.r
\name{eapr1}
\alias{eapr1}
\title{Convert given edu distribution into eapr within data.table}
\usage{
eapr1(prop)
}
\arguments{
\item{prop}{distribution data 'edu' (sorted) 'value' for single age-group}
}
\value{
eapr
}
\description{
This function returns EAPR for given education distribution
}
\keyword{read}
