% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funlinkcctext.r
\name{f.unlink.cc.text}
\alias{f.unlink.cc.text}
\title{delete specific file(s)}
\usage{
f.unlink.cc.text(cc, ipatt)
}
\arguments{
\item{cc}{region}

\item{ipatt}{pattern to find withing the folder}
}
\value{

}
\description{
This function finds and deletes file(s)
}
\examples{
load.x("filename")
}
\keyword{file}
\keyword{read}
