% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funlines.r
\name{funlines}
\alias{funlines}
\title{lines by scenarios (or any other variables)}
\usage{
funlines(res1, ivar, ititle, iby, iylab, ixlab, iregions = regions.all)
}
\arguments{
\item{res1}{data supplied with three columns scen Time value}

\item{ititle}{title}

\item{iby}{.BY from datatable group by}

\item{iylab}{ylab}

\item{ixlab}{xlab}

\item{iregions}{regions.all this is file saved in the mdepop}

\item{var1}{variable to plot}
}
\value{
a line graph
}
\description{
This function allows you to generate line graphs
}
\keyword{graph}
\keyword{line}
