% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwpppopas.r
\name{fwpppopA5G}
\alias{fwpppopA5G}
\title{Population past only for the moment by age and sex from WPP}
\usage{
fwpppopA5G(iwpp = "wpp2022")
}
\arguments{
\item{iwpp}{"wpp2022" as default}
}
\value{
popas region, Time, sex, agest, pop
}
\description{
This function extracts population from WPP and prepare it to match the msdem
}
\keyword{UN}
\keyword{population}
