% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popkc.r
\name{popkc}
\alias{popkc}
\title{Population size in MSDem}
\usage{
popkc(
  resX = res,
  period = NULL,
  agel = 0,
  ageu = "max",
  byage = FALSE,
  sex = NULL,
  bysex = FALSE,
  reg = NULL,
  byreg = FALSE,
  resi = NULL,
  byresi = FALSE
)
}
\arguments{
\item{period}{specific periods - default is NULL}

\item{agel}{lower limit - default is 0}

\item{ageu}{upper limit - default it 'max'}

\item{sex}{specific sex - default is NULL}

\item{bysex}{group by sex - default is FALSE}

\item{reg}{specific region(s)}

\item{byreg}{group by regions - default is FALSE}

\item{resi}{specific residence(s)}

\item{byresi}{group by residences - default is FALSE}

\item{res1}{data}
}
\value{
R list with five data.table 1) var_def 2) state_space 3) mig_dom 4) mig_int 5) results
}
\description{
This function allows you to calculate population size by sex or by regions using the MSDem output
}
\keyword{read}
