% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyr_ag}
\alias{pyr_ag}
\title{Population Pyramid in MSDem ??}
\usage{
pyr_ag(
  res1,
  ireg,
  ivar = "pop",
  iTime,
  iiscen = NULL,
  iscale = 1,
  facet.scale = "fixed",
  ititle.text = NULL,
  icol = "xxx"
)
}
\arguments{
\item{res1}{with four colums age, sex (male/female), edu (optional, will be aggregated), value (agest will be replaced by age; ivar will be replaced by value)}

\item{ireg}{region, this is needed for caption}

\item{ivar}{variable to be plotted (default pop)}

\item{iTime}{year, this is needed for caption}

\item{iiscen}{specific scenario, empirical (then no scenario is needed)}

\item{iscale}{unit of value (default 1 = '000, 1000 = millions, 1000000 - billions)}

\item{facet.scale}{facet scale (default fixed)}

\item{ititle.text}{full title text (default NULL)}
}
\value{
pyramid
}
\description{
This function generates a pyramid for a cntry x Time
}
\keyword{pyramid}
