% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R, R/pyredu.r
\name{pyredu}
\alias{pyredu}
\title{Population Pyramid in MSDem ??}
\usage{
pyredu(
  df1,
  Time,
  area,
  scen,
  popunit = "Thousands",
  caption = NULL,
  nmlegend,
  edu.nmlegend,
  ipal = 1
)

pyredu(
  df1,
  Time,
  area,
  scen,
  popunit = "Thousands",
  caption = NULL,
  nmlegend,
  edu.nmlegend,
  ipal = 1
)
}
\arguments{
\item{area}{region}

\item{scen}{specific scenario}

\item{popunit}{unit of value (default '000)}

\item{caption}{Full list else year area scen will be pasted}

\item{nmlegend}{legends for 'edu' else variable's unique values will be used}

\item{edu.nmlegend}{names of education categories}

\item{df}{with four colums age, sex (male/female), edu, value}

\item{year}{year}
}
\value{
pyramid

pyramid
}
\description{
This function generates a pyramid for a cntry x Time

This function generates a pyramid for a cntry x Time
}
\examples{
pyredu(df=df,year=2015,area="ARM",scen="Baseyear",nmlegend="Education")
pyredu(df=df,year=2015,area="ARM",scen="Baseyear",nmlegend="Education")
}
\keyword{pyramid}
