% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramidfacet.r
\name{pyredufacet}
\alias{pyredufacet}
\title{Population Pyramid in MSDem ??}
\usage{
pyredufacet(
  df1,
  year,
  area,
  scen,
  popunit = "Thousands",
  caption = NULL,
  nmlegend,
  ipal = 1,
  ixlab = "Age",
  iylab = "Population",
  female_nm = "female"
)
}
\arguments{
\item{df1}{with five colums age, sex (male/female), edu, varfacet,value}

\item{year}{year}

\item{area}{region}

\item{scen}{specific scenario}

\item{varfacet}{variable to be faceted}
}
\value{
pyramid
}
\description{
This function generates a pyramid for a cntry x Time
}
\examples{
pyredufacet(df1=df1,year=2015,area="ARM",scen="Baseyear",nmlegend="Education")
}
\keyword{pyramid}
