% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsaveVarCntry.r
\name{save.var.cntry.file}
\alias{save.var.cntry.file}
\title{save file}
\usage{
save.var.cntry.file(x, ivar)
}
\arguments{
\item{x}{value or df to be saved}

\item{ivar}{id (without region)}
}
\value{
df with 1
}
\description{
This function saves a given value to a file name
}
\examples{
save.var.cntry.file("value or df", "id without region")
}
\keyword{file}
\keyword{save}
