% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctocntry.r
\name{un_logit}
\alias{un_logit}
\title{unlogits}
\usage{
un_logit(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
unlogit of x
}
\description{
This transform back
}
\examples{
un_logit(1.2)
}
\keyword{logit}
\keyword{un}
