% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapr.r
\name{eapr}
\alias{eapr}
\title{Convert given edu distribution into eapr}
\usage{
eapr(prop)
}
\arguments{
\item{prop}{distribution data 'edu' (sorted) 'value' for single age-group}
}
\value{
eapr
}
\description{
This function returns EAPR for given education distribution
}
\keyword{read}
