% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpop.r
\name{mdpop}
\alias{mdpop}
\title{mdpop}
\usage{
mdpop(logit)
}
\value{
y unlogit value
}
\description{
This package allows you to work with mdpop and has the list of functions
}
\examples{
un_logit(1)
}
\keyword{logit}
