% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdpop.r
\name{mdpop}
\alias{mdpop}
\title{mdpop}
\usage{
mdpop(logit)
}
\arguments{
\item{cctoname}{xxx}

\item{eapr}{xxx}

\item{fread_all}{xxx}

\item{fsaveVarCntry}{xxx}

\item{funlinkcctext}{xxx}

\item{loadx}{xxx}

\item{cntocc}{xxx}

\item{popinterpkc}{xxx}

\item{popkc}{xxx}

\item{propkc}{xxx}

\item{pyramid}{xxx}

\item{pyredu}{xxx}

\item{un_logit}{xxx}
}
\value{
y unlogit value
}
\description{
This package allows you to work with mdpop and has the list of functions:
cctoname, eapr, fread_all, fsaveVarCntry, funlinkcctext, loadx, cntocc, popinterpkc, popkc, propkc, pyramid, pyredu
}
\examples{
mdpop(1)
}
\keyword{logit}
