% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyr_ag}
\alias{pyr_ag}
\title{Population Pyramid in MSDem ??}
\usage{
pyr_ag(
  res1,
  ireg,
  ivar = "pop",
  iTime,
  iiscen,
  iscale = 1,
  facet.scale = "fixed",
  ititle.text = NULL,
  icol = "xxx"
)
}
\arguments{
\item{res1}{with four colums age, sex (male/female), edu, value}

\item{Time}{year}

\item{region}{region}

\item{scen}{specific scenario}
}
\value{
pyramid
}
\description{
This function generates a pyramid for a cntry x Time
}
\keyword{pyramid}
