% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyredu}
\alias{pyredu}
\title{Population Pyramid in MSDem}
\usage{
pyredu(
  df,
  year,
  area,
  scen,
  popunit = "Thousands",
  caption = NULL,
  nmlegend,
  ipal = 1
)
}
\arguments{
\item{df}{with four colums age, sex (male/female), edu, value}

\item{year}{year}

\item{area}{region}

\item{scen}{specific scenario}
}
\value{
pyramid
}
\description{
This function generates a pyramid for a cntry x Time
}
\examples{
pyredu(df=df,year=2015,area="ARM",scen="Baseyear",nmlegend="Education")
}
\keyword{pyramid}
