% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R, R/pyredu.r
\name{pyredu}
\alias{pyredu}
\title{Population Pyramid in MSDem ??}
\usage{
pyredu(
  df1,
  Time,
  area,
  scen,
  popunit = "Thousands",
  caption = NULL,
  nmlegend,
  edu.nmlegend = NULL,
  limits = NULL,
  ipal = 1
)

pyredu(
  df1,
  Time,
  area,
  scen,
  popunit = "Thousands",
  caption = NULL,
  nmlegend,
  edu.nmlegend = NULL,
  limits = NULL,
  ipal = 1
)
}
\arguments{
\item{area}{region (will be used for caption, if caption is NULL)}

\item{scen}{specific scenario (will be used for caption, if caption is NULL)}

\item{popunit}{unit of value (default '000)}

\item{caption}{Full list else year, area, and scen will be pasted}

\item{nmlegend}{heading of legend}

\item{edu.nmlegend}{names of education categories NULL=takes names}

\item{ipal}{if numeric, it uses colors from brewer; else, you can supply the names of the colors manually}

\item{df}{with four colums age, sex (male/female), edu, value(e.g. population)}

\item{year}{year}
}
\value{
pyramid

pyramid
}
\description{
This function generates a pyramid for a cntry x Time

This function generates a pyramid for a cntry x Time
}
\examples{
pyredu(df=df,year=2015,area="ARM",scen="Baseyear",nmlegend="Education")
pyredu(df=df,year=2015,area="ARM",scen="Baseyear",nmlegend="Education",ipal=1)
}
\keyword{pyramid}
