% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un_logit.r
\name{un_logit}
\alias{un_logit}
\title{unlogit}
\usage{
un_logit(logit)
}
\arguments{
\item{logit}{numeric value}
}
\value{
y unlogit value
}
\description{
This function allows you to unlogit (back transform)
}
\examples{
un_logit(1)
}
\keyword{logit}
